refactor: move scene logic to module
